+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/audio/AudioMode.kt
@@ -25,10 +25,25 @@ enum class AudioMode(val value: Int) {
      */
     Stereo48K(3),
 
+    /**
+     * The [NoDevice] audio mode is obsolete, and is replaced by [AudioDeviceCapabilities.None]. To achieve the
+     * same functionality as NoDevice, pass AudioDeviceCapabilities.None into the AudioVideoConfiguration constructor
+     * instead, e.g. AudioVideoConfiguration(audioDeviceCapabilities = AudioDeviceCapabilities.None)
+     */
+    @Deprecated("To achieve the same functionality as NoDevice, pass AudioDeviceCapabilities.None into" +
+            " the AudioVideoConfiguration constructor instead, e.g." +
+            " AudioVideoConfiguration(audioDeviceCapabilities = AudioDeviceCapabilities.None)", level = DeprecationLevel.HIDDEN)
     NoDevice(4);
 
     companion object {
-        fun from(intValue: Int): AudioMode? = values().find { it.value == intValue }
+        fun from(intValue: Int): AudioMode? {
+            if (intValue == 4) {
+                // NoDevice cannot be instantiated since it is obsolete
+                return null
+            }
+            return values().find { it.value == intValue }
+        }
+
         fun from(intValue: Int, defaultAudioMode: AudioMode): AudioMode = from(intValue) ?: defaultAudioMode
     }
 }