+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/DefaultAudioClientController.kt
@@ -15,6 +15,7 @@ import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.analytics.MeetingStatsCollector
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.PrimaryMeetingPromotionObserver
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioDeviceCapabilities
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioMode
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioRecordingPresetOverride
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioStreamType
@@ -24,6 +25,7 @@ import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionStatus
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionStatusCode
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.xodee.client.audio.audioclient.AudioClient
+import com.xodee.client.audio.audioclient.AudioClient.AudioDeviceCapabilitiesInternal
 import com.xodee.client.audio.audioclient.AudioClient.AudioModeInternal
 import com.xodee.client.audio.audioclient.AudioClientSessionConfig
 import kotlinx.coroutines.CoroutineScope
@@ -68,7 +70,7 @@ class DefaultAudioClientController(
             AudioMode.Mono16K -> 16000
             AudioMode.Mono48K -> 48000
             AudioMode.Stereo48K -> 48000
-            AudioMode.NoDevice -> 16000
+            else -> 48000
         }
         audioClient.sendMessage(
             AudioClient.MESS_SET_IO_SAMPLE_RATE,
@@ -80,7 +82,7 @@ class DefaultAudioClientController(
             AudioMode.Mono16K -> AudioFormat.CHANNEL_OUT_MONO
             AudioMode.Mono48K -> AudioFormat.CHANNEL_OUT_MONO
             AudioMode.Stereo48K -> AudioFormat.CHANNEL_OUT_STEREO
-            AudioMode.NoDevice -> AudioFormat.CHANNEL_OUT_MONO
+            else -> AudioFormat.CHANNEL_OUT_STEREO
         }
         val spkMinBufSizeInSamples = AudioTrack.getMinBufferSize(
             nativeSR,
@@ -132,6 +134,7 @@ class DefaultAudioClientController(
         attendeeId: String,
         joinToken: String,
         audioMode: AudioMode,
+        audioDeviceCapabilities: AudioDeviceCapabilities,
         audioStreamType: AudioStreamType,
         audioRecordingPresetOverride: AudioRecordingPresetOverride,
         enableAudioRedundancy: Boolean
@@ -193,8 +196,10 @@ class DefaultAudioClientController(
                 AudioMode.Mono16K -> AudioModeInternal.MONO_16K
                 AudioMode.Mono48K -> AudioModeInternal.MONO_48K
                 AudioMode.Stereo48K -> AudioModeInternal.STEREO_48K
-                AudioMode.NoDevice -> AudioModeInternal.NO_DEVICE
+                else -> AudioModeInternal.STEREO_48K
             }
+
+            val audioDeviceCapabilitiesInternal = mapAudioDeviceCapabilitiesToInternal(audioDeviceCapabilities)
 
             var audioStreamTypeInternal = when (audioStreamType) {
                 AudioStreamType.VoiceCall -> AudioClient.AudioStreamType.VOICE_CALL
@@ -218,6 +223,7 @@ class DefaultAudioClientController(
                 audioFallbackUrl,
                 appInfo,
                 audioModeInternal,
+                audioDeviceCapabilitiesInternal,
                 audioStreamTypeInternal,
                 audioRecordingPresetInternal,
                 enableAudioRedundancy
@@ -335,5 +341,11 @@ class DefaultAudioClientController(
             return
         }
         audioClient.demoteFromPrimaryMeeting()
+    }
+
+    private fun mapAudioDeviceCapabilitiesToInternal(audioDeviceCapabilities: AudioDeviceCapabilities): AudioDeviceCapabilitiesInternal = when (audioDeviceCapabilities) {
+        AudioDeviceCapabilities.None -> AudioDeviceCapabilitiesInternal.NONE
+        AudioDeviceCapabilities.OutputOnly -> AudioDeviceCapabilitiesInternal.OUTPUT_ONLY
+        AudioDeviceCapabilities.InputAndOutput -> AudioDeviceCapabilitiesInternal.INPUT_AND_OUTPUT
     }
 }