+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/audio/AudioModeTest.kt
@@ -9,12 +9,12 @@ class AudioModeTest {
         Assert.assertEquals(AudioMode.from(1), AudioMode.Mono16K)
         Assert.assertEquals(AudioMode.from(2), AudioMode.Mono48K)
         Assert.assertEquals(AudioMode.from(3), AudioMode.Stereo48K)
-        Assert.assertEquals(AudioMode.from(4), AudioMode.NoDevice)
     }
 
     @Test
     fun `get enum value from invalid int returns null`() {
         Assert.assertNull(AudioMode.from(-1))
+        Assert.assertNull(AudioMode.from(4))
     }
 
     @Test